requirements.txt
discord.py~=1.7.0
dnspython==1.16.0
PyNaCl==1.3.0
async-timeout==3.0.1
python-dotenv~=0.17.0